Rehome Ring to own module
