+++ Data/Euclidean.hs
@@ -33,7 +33,8 @@ import Data.Complex (Complex(..))
 import Data.Int (Int, Int8, Int16, Int32, Int64)
 import Data.Maybe (isJust)
 import Data.Ratio (Ratio)
-import Data.Semiring (Semiring(..), Ring(..), (*), minus, isZero, Mod2)
+import Data.Semiring (Semiring(..), (*), isZero, Mod2)
+import Data.Ring (Ring(negate), minus)
 import Data.Word (Word, Word8, Word16, Word32, Word64)
 import Foreign.C.Types (CFloat, CDouble)
